+++ src/impl/dmx_hal.h
@@ -235,7 +235,7 @@ IRAM_ATTR uint32_t dmx_hal_get_rx_level(uart_hal_context_t *hal) {
  * @return The number of characters read.
  */
 IRAM_ATTR void dmx_hal_read_rxfifo(uart_hal_context_t *hal, uint8_t *buf,
-                                   size_t *size) {
+                                   int *size) {
   const size_t rxfifo_len = uart_ll_get_rxfifo_len(hal->dev);
   if (*size > rxfifo_len) {
     *size = rxfifo_len;
